More refactoring, type fixing, enable pyright strict for specific files
