+++ pyrightconfig.json
@@ -1,9 +1,100 @@
 {
   "include": [
-    "weave"
+    "weave/",
   ],
-  
   "exclude": [
-    "weave/frontend"
+    "docker_build.py",
+    "examples",
+    "integration_test",
+    "internal_tools",
+    "setup.py",
+    "weave-js",
+    "mypy_weave_plugin.py",
+    "weave/api.py",
+    "weave/artifact_fs.py",
+    "weave/artifact_local.py",
+    "weave/artifact_wandb.py",
+    "weave/arrow_util.py",
+    "weave/async_demo.py",
+    "weave/async_queue.py",
+    "weave/box.py",
+    "weave/cli.py",
+    "weave/client_interface.py",
+    "weave/codify.py",
+    "weave/compile_domain.py",
+    "weave/conftest.py",
+    "weave/context.py",
+    "weave/context_state.py",
+    "weave/decorator_class.py",
+    "weave/decorator_op.py",
+    "weave/decorator_type.py",
+    "weave/deploy",
+    "weave/derive_op.py",
+    "weave/dispatch.py",
+    "weave/ecosystem",
+    "weave/engine_trace.py",
+    "weave/execute.py",
+    "weave/file_base.py",
+    "weave/file_local.py",
+    "weave/filesystem.py",
+    "weave/gql_schema.py",
+    "weave/gql_to_weave.py",
+    "weave/infer_types.py",
+    "weave/io_service.py",
+    "weave/language_features",
+    "weave/logs.py",
+    "weave/mappers_arrow.py",
+    "weave/mappers_gql.py",
+    "weave/mappers_publisher.py",
+    "weave/mappers_python.py",
+    "weave/mappers_python_def.py",
+    "weave/monitoring",
+    "weave/node_ref.py",
+    "weave/op_def.py",
+    "weave/op_def_type.py",
+    "weave/ops_arrow",
+    "weave/ops_domain",
+    "weave/ops_primitives",
+    "weave/panels",
+    "weave/panel.py",
+    "weave/panel_util.py",
+    "weave/panels_py",
+    "weave/partial_object.py",
+    "weave/pyfunc_type_util.py",
+    "weave/registry_mem.py",
+    "weave/run_streamtable_span.py",
+    "weave/runfiles_wandb.py",
+    "weave/serialize.py",
+    "weave/serve_fastapi.py",
+    "weave/server.py",
+    "weave/server_error_handling.py",
+    "weave/show.py",
+    "weave/signal_handlers.py",
+    "weave/stream_data_interfaces.py",
+    "weave/storage.py",
+    "weave/syndata.py",
+    "weave/syndata_mon.py",
+    "weave/tests",
+    "weave/types_numpy.py",
+    "weave/usage_analytics.py",
+    "weave/util.py",
+    "weave/wandb_api.py",
+    "weave/wandb_client_api.py",
+    "weave/wandb_file_manager.py",
+    "weave/wandb_interface",
+    "weave/wandb_util.py",
+    "weave/weave_inspector.py",
+    "weave/weave_internal.py",
+    "weave/weave_server.py",
+    "weave/weave_types.py",
+    "weave/weaveflow",
+    "weave/weave_http.py",
+    "weave/weave_pydantic.py",
+    "weave/weavify.py"
+  ],
+  "strict": [
+    "weave/ref_base.py",
+    "weave/graph_client.py",
+    "weave/graph_client_local.py"
   ]
 }
