+++ weave/op_def.py
@@ -19,7 +19,7 @@ from . import engine_trace
 from . import memo
 from . import weavify
 from . import eager
-from . import run
+from .run import Run
 from . import graph_client_context
 
 from .language_features.tagging import (
@@ -606,8 +606,7 @@ class OpDef:
     def op_def_is_auto_tag_handling_arrow_op(self) -> bool:
         return isinstance(self, AutoTagHandlingArrowOpDef)
 
-    # TODO: Should be generic Run protocol, but need to update graph_client type first.
-    def runs(self) -> eager.WeaveIter["RunStreamTableSpan"]:
+    def runs(self) -> eager.WeaveIter[Run]:
         client = graph_client_context.require_graph_client()
         return client.op_runs(self)
