+++ weave/ops_primitives/weave_api.py
@@ -369,7 +369,7 @@ def _merge(name) -> str:
                 "Cannot merge into artifact without version"
             )
 
-        ref = storage._direct_save(
+        ref = storage.direct_save(
             obj=head_ref.get(),
             name=to_uri.name,
             source_branch_name=shared_branch_name,
