+++ weave/run_streamtable_span.py
@@ -1,6 +1,7 @@
 import typing
 
 from .urls import BROWSE3_PATH
+from .run import Run
 from . import uris
 from . import artifact_wandb
 from . import stream_data_interfaces
@@ -32,7 +33,7 @@ class RunStreamTableSpan:
     @property
     def ui_url(self) -> str:
         gc = graph_client_context.require_graph_client()
-        return f"http://localhost:3000/{BROWSE3_PATH}/{gc.entity_name}/{gc.project_name}/calls/{self.id}"
+        return gc.run_ui_url(self)
 
     @property
     def op_ref(self) -> typing.Optional[artifact_wandb.WandbArtifactRef]:
@@ -68,15 +69,15 @@ class RunStreamTableSpan:
     def output(self) -> typing.Any:
         if self.status_code != "SUCCESS":
             return None
-        output_dict = self._attrs.get("output")
-        if isinstance(output_dict, dict):
-            keys = output_dict.get("_keys")
+        out = self._attrs.get("output")
+        if isinstance(out, dict):
+            keys = out.get("_keys")
             if keys is None:
-                keys = [k for k in output_dict.keys() if output_dict[k] != None]
-            output = {k: output_dict.get(k) for k in keys}
+                keys = [k for k in out.keys() if out[k] != None]
+            output = {k: out.get(k) for k in keys}
             if "_result" in output:
                 return output["_result"]
-        return output
+        return out
 
     @property
     def parent_id(self) -> typing.Optional[str]:
@@ -90,7 +91,7 @@ class RunStreamTableSpan:
         client = graph_client_context.require_graph_client()
         return client.run_feedback(self.id)
 
-    def parent(self) -> typing.Optional["RunStreamTableSpan"]:
+    def parent(self) -> typing.Optional["Run"]:
         client = graph_client_context.require_graph_client()
         if self.parent_id is None:
             return None
