+++ weave/tests/test_publish_flow.py
@@ -125,6 +125,7 @@ In summary, there will be 22 tests:
    * -2 cases that are not possible
 """
 
+
 # Persist tests
 def test_persist_to_remote_with_commit_hash(user_by_api_key_in_env):
     data = ["test_data"]
@@ -188,7 +189,7 @@ def test_persist_to_local_with_commit_hash(user_by_api_key_in_env):
     expected_commit_hash = "b11179315e19b4207282"
     branch_name = None
 
-    p_ref = storage._direct_save(
+    p_ref = storage.direct_save(
         obj=data,
         name=target_artifact_name,
         branch_name=branch_name,
@@ -215,7 +216,7 @@ def test_persist_to_local_with_branch(user_by_api_key_in_env):
     expected_commit_hash = "b11179315e19b4207282"
     branch_name = "my_branch"
 
-    p_ref = storage._direct_save(
+    p_ref = storage.direct_save(
         obj=data,
         name=target_artifact_name,
         branch_name=branch_name,
@@ -241,14 +242,14 @@ def test_persist_to_local_with_branch_and_branchpoint(user_by_api_key_in_env):
     source_commit_hash = "741eb73e40d3b38b046b"
 
     # First, save a local artifact with a branch
-    p_ref = storage._direct_save(
+    p_ref = storage.direct_save(
         obj=["initial_data"],
         name=target_artifact_name,
         branch_name=source_branch_name,
     )
 
     # Then, save a new local artifact with a branchpoint
-    new_p_ref = storage._direct_save(
+    new_p_ref = storage.direct_save(
         obj=data,
         name=target_artifact_name,
         branch_name=branch_name,
@@ -446,7 +447,7 @@ def test_mutate_local_with_commit_hash(user_by_api_key_in_env, new_branch_name=N
     new_data = ["test_data_2"]
     new_commit_hash = "61f78c8877df22942d23"
 
-    p_ref = storage._direct_save(
+    p_ref = storage.direct_save(
         obj=data,
         name=target_artifact_name,
         branch_name=branch_name,
@@ -556,7 +557,7 @@ def test_mutate_local_with_branch(user_by_api_key_in_env, new_branch_name=None):
     new_data = ["test_data_2"]
     new_commit_hash = "61f78c8877df22942d23"
 
-    p_ref = storage._direct_save(
+    p_ref = storage.direct_save(
         obj=data,
         name=target_artifact_name,
         branch_name=branch_name,
